Rename MethodHandleHelper to MethodHandleResolver
This is required to avoid name conflicts with OpenJDK tests which
inject a java.lang.invoke.MethodHandleHelper class during the test.

Related: #10857,#10096
