+++ jcl/src/java.base/share/classes/java/lang/invoke/IndirectHandle.java
@@ -26,8 +26,8 @@ package java.lang.invoke;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 
-import static java.lang.invoke.MethodHandleHelper.UNSAFE;
-import static java.lang.invoke.MethodHandleHelper.getJ9ClassFromClass;
+import static java.lang.invoke.MethodHandleResolver.UNSAFE;
+import static java.lang.invoke.MethodHandleResolver.getJ9ClassFromClass;
 
 abstract class IndirectHandle extends PrimitiveHandle {
 	IndirectHandle(MethodType type, Class<?> referenceClass, String name, byte kind, int modifiers) {
