+++ jcl/src/java.base/share/classes/java/lang/invoke/InterfaceHandle.java
@@ -24,7 +24,7 @@ package java.lang.invoke;
 
 import java.lang.reflect.Method;
 
-import static java.lang.invoke.MethodHandleHelper.getJ9ClassFromClass;
+import static java.lang.invoke.MethodHandleResolver.getJ9ClassFromClass;
 
 /*
  * InterfaceHandle is a MethodHandle that does interface dispatch
