+++ jcl/src/java.base/share/classes/java/lang/invoke/PrimitiveHandle.java
@@ -34,8 +34,8 @@ import com.ibm.jit.JITHelpers;
 import java.util.List;
 /*[ENDIF] Java15 */
 
-import static java.lang.invoke.MethodHandleHelper.UNSAFE;
-import static java.lang.invoke.MethodHandleHelper.JITHELPERS;
+import static java.lang.invoke.MethodHandleResolver.UNSAFE;
+import static java.lang.invoke.MethodHandleResolver.JITHELPERS;
 
 /**
  * PrimitiveHandle is a subclass of MethodHandle used for grouping MethodHandles that directly refer a Java-level method.
