+++ runtime/jcl/exports.cmake
@@ -362,10 +362,10 @@ omr_add_exports(jclse
 	Java_java_lang_Thread_suspendImpl
 	Java_java_lang_Thread_yield
 	Java_java_lang_invoke_InterfaceHandle_registerNatives
-	Java_java_lang_invoke_MethodHandleHelper_getCPClassNameAt
-	Java_java_lang_invoke_MethodHandleHelper_getCPMethodHandleAt
-	Java_java_lang_invoke_MethodHandleHelper_getCPMethodTypeAt
-	Java_java_lang_invoke_MethodHandleHelper_getCPTypeAt
+	Java_java_lang_invoke_MethodHandleResolver_getCPClassNameAt
+	Java_java_lang_invoke_MethodHandleResolver_getCPMethodHandleAt
+	Java_java_lang_invoke_MethodHandleResolver_getCPMethodTypeAt
+	Java_java_lang_invoke_MethodHandleResolver_getCPTypeAt
 	Java_java_lang_invoke_MethodHandle_invoke
 	Java_java_lang_invoke_MethodHandle_invokeExact
 	Java_java_lang_invoke_MethodHandle_requestCustomThunkFromJit
@@ -574,7 +574,7 @@ if(NOT JAVA_SPEC_VERSION LESS 11)
 	omr_add_exports(jclse
 		Java_java_lang_Class_getNestHostImpl
 		Java_java_lang_Class_getNestMembersImpl
-		Java_java_lang_invoke_MethodHandleHelper_getCPConstantDynamicAt
+		Java_java_lang_invoke_MethodHandleResolver_getCPConstantDynamicAt
 	)
 endif()
