+++ runtime/jcl/uma/se7_exports.xml
@@ -54,10 +54,10 @@
 	<export name="Java_java_lang_invoke_MethodHandle_vmRefFieldOffset" />
 	<export name="Java_java_lang_invoke_MethodHandle_invoke" />
 	<export name="Java_java_lang_invoke_MethodHandle_invokeExact" />
-	<export name="Java_java_lang_invoke_MethodHandleHelper_getCPTypeAt" />
-	<export name="Java_java_lang_invoke_MethodHandleHelper_getCPMethodTypeAt" />
-	<export name="Java_java_lang_invoke_MethodHandleHelper_getCPMethodHandleAt" />
-	<export name="Java_java_lang_invoke_MethodHandleHelper_getCPClassNameAt" />
+	<export name="Java_java_lang_invoke_MethodHandleResolver_getCPTypeAt" />
+	<export name="Java_java_lang_invoke_MethodHandleResolver_getCPMethodTypeAt" />
+	<export name="Java_java_lang_invoke_MethodHandleResolver_getCPMethodHandleAt" />
+	<export name="Java_java_lang_invoke_MethodHandleResolver_getCPClassNameAt" />
 	<export name="Java_java_lang_invoke_ThunkTuple_registerNatives" />
 	<export name="Java_java_lang_invoke_InterfaceHandle_registerNatives" />
 	<export name="Java_java_lang_invoke_MethodType_makeTenured" />
